0.2.6 even more merging PRs
